Ex 2.10 -  components in separate files
