+++ vite/phonebook/src/App.jsx
@@ -1,41 +1,7 @@
 import { useState } from 'react'
-const Input = ({ inputText, inputValue, onChange }) => {
-  return (
-    <div>
-      {inputText}: <input value={inputValue} onChange={onChange} />
-    </div>
-  )
-}
-
-const Filter = ({ searchText, handleChange }) => {
-  return (
-    <>
-      < Input inputText="filter shown with " inputValue={searchText} onChange={handleChange} />
-    </>
-  )
-}
-
-const PersonForm = ({ newName, newNumber, handleNameChange, handleNumberChange, submitAction }) => {
-  return (
-    <>
-      <form onSubmit={submitAction}>
-        <Input inputText="name" inputValue={newName} onChange={handleNameChange} />
-        <Input inputText="number" inputValue={newNumber} onChange={handleNumberChange} />
-        <div>
-          <button type="submit">add</button>
-        </div>
-      </form>
-    </>
-  )
-}
-
-const Persons = ({ persons }) => {
-  return (
-    <>
-      {persons.map(person => <div key={person.name}>{person.name} {person.number} </div>)}
-    </>
-  )
-}
+import Filter from './components/filter'
+import PersonForm from './components/personform'
+import Persons from './components/persons'
 
 const App = () => {
   const [persons, setPersons] = useState([
