+++ vite/phonebook/src/components/input.jsx
@@ -1,0 +1,8 @@
+const Input = ({ inputText, inputValue, onChange }) => {
+    return (
+        <div>
+            {inputText}: <input value={inputValue} onChange={onChange} />
+        </div>
+    )
+}
+export default Input
