+++ vite/phonebook/src/components/persons.jsx
@@ -1,0 +1,9 @@
+const Persons = ({ persons }) => {
+    return (
+        <>
+            {persons.map(person => <div key={person.name}>{person.name} {person.number} </div>)}
+        </>
+    )
+}
+
+export default Persons
